refactor

+++ src/app/api/savings/utils/del.ts
@@ -1,34 +1,15 @@
-import { authOptions } from "@/lib/auth";
-import { SessionType, Transaction as TransactionType } from "@/lib/types";
 import Savings from "@/models/Savings";
 import Transaction from "@/models/Transaction";
-import User from "@/models/User";
-import { getServerSession } from "next-auth";
-import { NextRequest } from "next/server";
+import { auth } from "@clerk/nextjs/server";
+import { NextRequest, NextResponse } from "next/server";
 
 export async function del(request: NextRequest) {
-	const session: SessionType = await getServerSession(authOptions);
+	const { userId } = auth();
 
-	if (!session) {
-		return new Response("unauthorized");
-	}
-
-	const userId = session.user.id;
+	if (!userId)
+		return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
 	const body = await request.json();
-
-	const transactions: TransactionType[] = await Transaction.find({
-		typeId: body._id,
-	});
-
-	const transactionIds = transactions.map((transaction) => transaction._id);
-
-	await User.findByIdAndUpdate(userId, {
-		$pull: {
-			transactionIds: { $in: transactionIds },
-			savingsIds: body._id,
-		},
-	});
 
 	await Transaction.deleteMany({ typeId: body._id });
